add core serverless and pass easy test
