+++ parsl/executors/taskvine/utils.py
@@ -78,7 +78,8 @@ class ParslFileToVine:
 
 
 def run_parsl_function(map_file, function_file, argument_file, result_file):
-    # This function assumes that parsl module is importable on remote nodes
-    # and executes function after reading relevant files.
+    """
+    Wrapper function to deploy with FunctionCall as serverless tasks.
+    """
     from parsl.executors.taskvine.exec_parsl_function import run
     run(map_file, function_file, argument_file, result_file)
