+++ parsl/tests/configs/taskvine_ex.py
@@ -9,5 +9,5 @@ from parsl.data_provider.file_noop import NoOpFileStaging
 
 def fresh_config():
     return Config(executors=[TaskVineExecutor(manager_config=TaskVineManagerConfig(port=9000),
-                                              use_factory=True,
+                                              worker_launch_method='factory',
                                               storage_access=[FTPInTaskStaging(), HTTPInTaskStaging(), NoOpFileStaging()])])
