+++ parsl/tests/scaling_tests/vineex_local.py
@@ -5,7 +5,7 @@ from parsl.providers import LocalProvider
 
 config = Config(
     executors=[TaskVineExecutor(label='VineExec',
-                                use_factory=True,
+                                worker_launch_method='factory',
                                 manager_config=TaskVineManagerConfig(port=50055),
         )]
 )
